+++ moonshine_version.h
@@ -3,6 +3,6 @@
 
 #define MOONSHINE_NAME "Moonshine"
 #define MOONSHINE_VERSION "0.1.0-research-preview"
-#define MOONSHINE_REPOSITORY_SLUG "moonshine-k3"
+#define MOONSHINE_REPOSITORY_SLUG "moonshine"
 
 #endif